backend/requirements.txt
fastapi
uvicorn
motor
langgraph
langchain-core
langchain-google-genai
pydantic-settings
python-dotenv
google-genai
python-jose[cryptography]
supabase
httpx